move getAllScanResults to shared
